+++ sentinel/src/main/kotlin/com/infinum/sentinel/domain/crash/monitor/models/CrashMonitorParameters.kt
@@ -1,0 +1,8 @@
+package com.infinum.sentinel.domain.crash.monitor.models
+
+import com.infinum.sentinel.data.models.local.CrashMonitorEntity
+import com.infinum.sentinel.domain.shared.base.BaseParameters
+
+internal data class CrashMonitorParameters(
+    val entity: CrashMonitorEntity? = null
+) : BaseParameters