+++ sentinel/src/main/kotlin/com/infinum/sentinel/ui/crash/notification/NotificationFactory.kt
@@ -1,0 +1,10 @@
+package com.infinum.sentinel.ui.crash.notification
+
+import com.infinum.sentinel.data.models.local.CrashEntity
+
+internal interface NotificationFactory {
+
+    fun showCrash(applicationName: String, id: Long, entity: CrashEntity)
+
+    fun showAnr(applicationName: String, id: Long, entity: CrashEntity)
+}