+++ sentinel/src/main/kotlin/com/infinum/sentinel/ui/main/SentinelFragment.kt
@@ -9,6 +9,7 @@ import com.google.android.material.shape.MaterialShapeDrawable
 import com.google.android.material.shape.ShapeAppearanceModel
 import com.infinum.sentinel.R
 import com.infinum.sentinel.databinding.SentinelFragmentBinding
+import com.infinum.sentinel.extensions.shareText
 import com.infinum.sentinel.extensions.toCradleDrawable
 import com.infinum.sentinel.ui.main.application.ApplicationFragment
 import com.infinum.sentinel.ui.main.device.DeviceFragment
@@ -27,8 +28,6 @@ internal class SentinelFragment : BaseFragment<SentinelState, SentinelEvent>(R.l
 
     companion object {
         const val TAG: String = "SentinelFragment"
-
-        private const val SHARE_MIME_TYPE = "text/plain"
     }
 
     override val viewModel: SentinelViewModel by viewModel()
@@ -63,10 +62,7 @@ internal class SentinelFragment : BaseFragment<SentinelState, SentinelEvent>(R.l
     override fun onEvent(event: SentinelEvent) =
         when (event) {
             is SentinelEvent.Formatted -> ShareCompat.IntentBuilder(requireActivity())
-                .setChooserTitle(R.string.sentinel_name)
-                .setType(SHARE_MIME_TYPE)
-                .setText(event.value)
-                .startChooser()
+                .shareText(event.value)
         }
 
     private fun setupToolbar() {
@@ -77,9 +73,7 @@ internal class SentinelFragment : BaseFragment<SentinelState, SentinelEvent>(R.l
             }
             toolbar.setOnMenuItemClickListener { menuItem ->
                 when (menuItem.itemId) {
-                    R.id.share -> {
-                        viewModel.formatData()
-                    }
+                    R.id.share -> viewModel.formatData()
                 }
                 true
             }