+++ sentinel/src/main/kotlin/com/infinum/sentinel/ui/settings/SettingsViewModel.kt
@@ -1,12 +1,16 @@
 package com.infinum.sentinel.ui.settings
 
 import com.infinum.sentinel.data.models.local.BundleMonitorEntity
+import com.infinum.sentinel.data.models.local.CrashMonitorEntity
 import com.infinum.sentinel.data.models.local.FormatEntity
 import com.infinum.sentinel.data.models.local.TriggerEntity
 import com.infinum.sentinel.domain.Repositories
 import com.infinum.sentinel.domain.bundle.monitor.models.BundleMonitorParameters
+import com.infinum.sentinel.domain.crash.monitor.models.CrashMonitorParameters
 import com.infinum.sentinel.domain.formats.models.FormatsParameters
 import com.infinum.sentinel.domain.triggers.models.TriggerParameters
+import com.infinum.sentinel.ui.crash.anr.SentinelAnrObserver
+import com.infinum.sentinel.ui.crash.handler.SentinelExceptionHandler
 import com.infinum.sentinel.ui.shared.base.BaseChildViewModel
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.flowOn
@@ -14,7 +18,10 @@ import kotlinx.coroutines.flow.flowOn
 internal class SettingsViewModel(
     private val triggers: Repositories.Triggers,
     private val formats: Repositories.Formats,
-    private val bundleMonitor: Repositories.BundleMonitor
+    private val bundleMonitor: Repositories.BundleMonitor,
+    private val crashMonitor: Repositories.CrashMonitor,
+    private val exceptionHandler: SentinelExceptionHandler,
+    private val anrObserver: SentinelAnrObserver,
 ) : BaseChildViewModel<Nothing, SettingsEvent>() {
 
     override fun data() {
@@ -37,6 +44,13 @@ internal class SettingsViewModel(
                 .flowOn(runningDispatchers)
                 .collectLatest {
                     emitEvent(SettingsEvent.BundleMonitorChanged(value = it))
+                }
+        }
+        launch {
+            crashMonitor.load(CrashMonitorParameters())
+                .flowOn(runningDispatchers)
+                .collectLatest {
+                    emitEvent(SettingsEvent.CrashMonitorChanged(value = it))
                 }
         }
     }
@@ -71,6 +85,27 @@ internal class SettingsViewModel(
                         entity = entity
                     )
                 )
+            }
+        }
+
+    fun updateCrashMonitor(entity: CrashMonitorEntity) =
+        launch {
+            io {
+                crashMonitor.save(
+                    CrashMonitorParameters(
+                        entity = entity
+                    )
+                )
+                if (entity.notifyExceptions) {
+                    exceptionHandler.start()
+                } else {
+                    exceptionHandler.stop()
+                }
+                if (entity.notifyAnrs) {
+                    anrObserver.start()
+                } else {
+                    anrObserver.stop()
+                }
             }
         }
 }